PR review changes

+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles.Tests/Lib/Helpers.cs
@@ -13,9 +13,9 @@ internal static class Helpers
     /// <summary>
     /// Test credentials for docker server.
     /// </summary>
-    private static string _dockerAddress = "localhost";
-    private static string _dockerUsername = "foo";
-    private static string _dockerPassword = "pass";
+    readonly static string _dockerAddress = "localhost";
+    readonly static string _dockerUsername = "foo";
+    readonly static string _dockerPassword = "pass";
 
     internal static Connection GetSftpConnection()
     {
@@ -71,18 +71,11 @@ internal static class Helpers
         {
             if ((file.Name != ".") && (file.Name != ".."))
             {
-                if (file.IsDirectory)
-                {
-                    DeleteDirectory(client, file.FullName);
-                }
-                else
-                {
-                    client.DeleteFile(file.FullName);
-                }
+                if (file.IsDirectory) DeleteDirectory(client, file.FullName);
+                else client.DeleteFile(file.FullName);
             }
         }
-        if (client.Exists(dir))
-            client.DeleteDirectory(dir);
+        if (client.Exists(dir)) client.DeleteDirectory(dir);
     }
 
     internal static void UploadTestFiles(List<string> paths, string destination, string to = null)
